+++ test/verify_test.dart
@@ -63,5 +63,26 @@ void main() {
           ]),
           2);
     });
+
+    test('isTrueMinimumNumberOfTimes', () {
+      final int left = 2;
+      final int right1 = 1 + 1;
+      final int right2 = 2 + 1;
+
+      expect(isTrueMinimumNumberOfTimes([true, false, false]), true);
+      expect(isTrueMinimumNumberOfTimes([true, true], minimumCount: 1), true);
+      expect(isTrueMinimumNumberOfTimes([true, false, true], minimumCount: 2),
+          true);
+      expect(isTrueMinimumNumberOfTimes([true, false, true], minimumCount: 1),
+          true);
+      expect(
+          isTrueMinimumNumberOfTimes([left == right1, false, true, true, false],
+              minimumCount: 3),
+          true);
+      expect(
+          isTrueMinimumNumberOfTimes([left == right2, false, true, true, false],
+              minimumCount: 3),
+          false);
+    });
   });
 }