+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `arrGroupByMaxCount` method
 - Add `numRandom` method
 - Add `len` method
+- Add `isTrueMinimumNumberOfTimes` method
 
 ## 1.0.0 (2024-10-19)